test(discord): make lifecycle shutdown ordering deterministic (#3145)
